feat: api 파일 하나로 통합 및 API_BASE_URL 생성
