--- src/components/main/api/linksApi.js
@@ -1,5 +1,0 @@
-export async function getLinks(id) {
-  const API_BASE_URL_LINKS = `https://bootcamp-api.codeit.kr/api/users/1/links${ id ? `?folderId=${id}` : " "}`
-  const response = await fetch(API_BASE_URL_LINKS)
-  return await response.json();
-}
